Expose `unquote_string` from `edgeql-parser`
